FE-7: feat(quiz) - extract optionComponent to dedicated file
