+++ web/src/components/quiz/questionComponent.tsx
@@ -1,25 +1,10 @@
 import { Box, HStack, Text, VStack } from '@chakra-ui/react';
 import { FC } from 'react';
-import Answer from "../../model/answer";
+import Option from "./optionComponent";
 import Question from "../../model/question";
-
-interface optionProps {
-    answer: Answer
-}
 
 interface questionProp {
     question: Question
-}
-
-const Option: FC<optionProps> = ({answer}) => {
-    return (
-        <Box>
-            <HStack>
-                <h3>{answer.text}</h3>
-                <Text>{answer.text}</Text>
-            </HStack>
-        </Box>
-    )
 }
 
 const QuestionComponent: FC<questionProp> =  ({question}) => {
